+++ gqt_cleanup/gqt_cleanup_suite_test.go
@@ -77,7 +77,10 @@ var _ = AfterEach(func() {
 		Expect(os.RemoveAll(containerdRunDir)).To(Succeed())
 	}
 
-	Expect(os.RemoveAll(config.TmpDir)).To(Succeed())
+	// Windows worker is not containerised and therefore the test needs to take care to delete the temporary folder
+	if runtime.GOOS == "windows" {
+		Expect(os.RemoveAll(config.TmpDir)).To(Succeed())
+	}
 })
 
 func terminateContainerd() {
